Refactor weather widget rendering and extract it to separate class
